# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Portal de compras - Farmácias Associadas",
        "version": "1.0.0"
    },
    "paths": {
...
        "/programs/{id}/connection": {
            "post": {
                "tags": [
                    "Programs"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": "2"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConnectionCreatorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "example": "Conexão criada com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "example": "Mensagem de erro"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/programs/{id}/connection/{connection}": {
            "put": {
                "tags": [
                    "Programs"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": "2"
                    },
                    {
                        "name": "connection",
                        "in": "path",
                        "required": true,
                        "example": "2"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConnectionUpdatorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "example": "Programa Conexão atualizado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "example": "Mensagem de erro"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ConnectionCreatorRequest": {
                "title": "DistributorConnection form request",
                "required": [
                    "isFtpActive",
                    "transferMode",
                    "host",
                    "sendDirectory",
                    "login",
                    "password",
                    "returnDirectory"
                ],
                "properties": {
                    "distributorId": {
                        "type": "boolean",
                        "example": "1"
                    },
                    "isFtpActive": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "transferMode": {
                        "type": "string",
                        "example": "ASCI"
                    },
                    "host": {
                        "type": "string",
                        "example": "127.0.0.1"
                    },
                    "sendDirectory": {
                        "type": "string",
                        "example": "/var/www"
                    },
                    "login": {
                        "type": "string",
                        "example": "teste"
                    },
                    "password": {
                        "type": "string",
                        "example": "123456"
                    },
                    "returnDirectory": {
                        "type": "string",
                        "example": "/var/www"
                    }
                },
                "type": "object"
            },
            "ConnectionUpdatorRequest": {
                "required": [
                    "isFtpActive",
                    "transferMode",
                    "host",
                    "sendDirectory",
                    "login",
                    "password",
                    "returnDirectory"
                ],
                "properties": {
                    "isFtpActive": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "transferMode": {
                        "type": "string",
                        "example": "ASCI"
                    },
                    "host": {
                        "type": "string",
                        "example": "127.0.0.1"
                    },
                    "sendDirectory": {
                        "type": "string",
                        "example": "/var/www"
                    },
                    "login": {
                        "type": "string",
                        "example": "teste"
                    },
                    "password": {
                        "type": "string",
                        "example": "123456"
                    },
                    "returnDirectory": {
                        "type": "string",
                        "example": "/var/www"
                    }
                },
                "type": "object"
            },
...
            "ValidationResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            },
                            "example": "Campo teste é obrigatório"
                        }
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}